+++ tests/native-test.js
@@ -14,6 +14,8 @@ module.exports = async function ({
   updates,
   updateAdapter,
 }) {
+  if (process.env.TABLELIGHT_TEST_SCENARIO === 'passives-dm')
+    return require('./passives-dm-native')({ app, controller, getState, store });
   if (process.env.TABLELIGHT_TEST_SCENARIO === 'passives')
     return require('./passives-native')({
       app,